pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rapidapi-helper"
version = "0.1.0"
description = "Учебный набор инструкций и примеров для работы с RapidAPI через ИИ-ассистентов"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "RapidAPI Helper Contributors" }]
dependencies = [
    "requests>=2.31",
    "requests-cache>=1.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pymarkdownlnt>=0.9",
]

[project.scripts]
rapidapi-new-connector = "scripts.new_connector:main"

[tool.setuptools.packages.find]
include = ["examples*"]
exclude = ["tests*", "docs*", "connectors*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
